+++ src/IO.Swagger.V1RC03/Controllers/AssetAdministrationShellBasicDiscoveryAPI.cs
@@ -17,7 +17,6 @@ using System.ComponentModel.DataAnnotations;
 using IO.Swagger.V1RC03.Attributes;
 
 using Microsoft.AspNetCore.Authorization;
-using IO.Swagger.V1RC03.Models;
 using IO.Swagger.V1RC03.ApiModel;
