+++ src/IO.Swagger.V1RC03/Controllers/ISubmodelRegistryAPI.cs
@@ -12,7 +12,7 @@ using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using System.ComponentModel.DataAnnotations;
-using IO.Swagger.V1RC03.Models;
+using IO.Swagger.V1RC03.ApiModel;
 
 namespace IO.Swagger.V1RC03.Controllers
 {
