--- src/IO.Swagger.V1RC03/Models/Endpoint.cs
@@ -1,137 +1,0 @@
-/*
- * DotAAS Part 2 | HTTP/REST | Entire API Collection
- *
- * The entire API collection as part of Details of the Asset Administration Shell Part 2
- *
- * OpenAPI spec version: V1.0RC03
- * 
- * Generated by: https://github.com/swagger-api/swagger-codegen.git
- */
-using System;
-using System.Linq;
-using System.IO;
-using System.Text;
-using System.Collections;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.ComponentModel.DataAnnotations;
-using System.Runtime.Serialization;
-using Newtonsoft.Json;
-
-namespace IO.Swagger.V1RC03.Models
-{ 
-    /// <summary>
-    /// 
-    /// </summary>
-    [DataContract]
-    public partial class Endpoint : IEquatable<Endpoint>
-    { 
-        /// <summary>
-        /// Gets or Sets _Interface
-        /// </summary>
-        [Required]
-
-        [DataMember(Name="interface")]
-        public string _Interface { get; set; }
-
-        /// <summary>
-        /// Gets or Sets ProtocolInformation
-        /// </summary>
-        [Required]
-
-        [DataMember(Name="protocolInformation")]
-        public ProtocolInformation ProtocolInformation { get; set; }
-
-        /// <summary>
-        /// Returns the string presentation of the object
-        /// </summary>
-        /// <returns>String presentation of the object</returns>
-        public override string ToString()
-        {
-            var sb = new StringBuilder();
-            sb.Append("class Endpoint {\n");
-            sb.Append("  _Interface: ").Append(_Interface).Append("\n");
-            sb.Append("  ProtocolInformation: ").Append(ProtocolInformation).Append("\n");
-            sb.Append("}\n");
-            return sb.ToString();
-        }
-
-        /// <summary>
-        /// Returns the JSON string presentation of the object
-        /// </summary>
-        /// <returns>JSON string presentation of the object</returns>
-        public string ToJson()
-        {
-            return JsonConvert.SerializeObject(this, Formatting.Indented);
-        }
-
-        /// <summary>
-        /// Returns true if objects are equal
-        /// </summary>
-        /// <param name="obj">Object to be compared</param>
-        /// <returns>Boolean</returns>
-        public override bool Equals(object obj)
-        {
-            if (ReferenceEquals(null, obj)) return false;
-            if (ReferenceEquals(this, obj)) return true;
-            return obj.GetType() == GetType() && Equals((Endpoint)obj);
-        }
-
-        /// <summary>
-        /// Returns true if Endpoint instances are equal
-        /// </summary>
-        /// <param name="other">Instance of Endpoint to be compared</param>
-        /// <returns>Boolean</returns>
-        public bool Equals(Endpoint other)
-        {
-            if (ReferenceEquals(null, other)) return false;
-            if (ReferenceEquals(this, other)) return true;
-
-            return 
-                (
-                    _Interface == other._Interface ||
-                    _Interface != null &&
-                    _Interface.Equals(other._Interface)
-                ) && 
-                (
-                    ProtocolInformation == other.ProtocolInformation ||
-                    ProtocolInformation != null &&
-                    ProtocolInformation.Equals(other.ProtocolInformation)
-                );
-        }
-
-        /// <summary>
-        /// Gets the hash code
-        /// </summary>
-        /// <returns>Hash code</returns>
-        public override int GetHashCode()
-        {
-            unchecked // Overflow is fine, just wrap
-            {
-                var hashCode = 41;
-                // Suitable nullity checks etc, of course :)
-                    if (_Interface != null)
-                    hashCode = hashCode * 59 + _Interface.GetHashCode();
-                    if (ProtocolInformation != null)
-                    hashCode = hashCode * 59 + ProtocolInformation.GetHashCode();
-                return hashCode;
-            }
-        }
-
-        #region Operators
-        #pragma warning disable 1591
-
-        public static bool operator ==(Endpoint left, Endpoint right)
-        {
-            return Equals(left, right);
-        }
-
-        public static bool operator !=(Endpoint left, Endpoint right)
-        {
-            return !Equals(left, right);
-        }
-
-        #pragma warning restore 1591
-        #endregion Operators
-    }
-}
