--- src/IO.Swagger.V1RC03/Models/PackageDescription.cs
@@ -1,135 +1,0 @@
-/*
- * DotAAS Part 2 | HTTP/REST | Entire API Collection
- *
- * The entire API collection as part of Details of the Asset Administration Shell Part 2
- *
- * OpenAPI spec version: V1.0RC03
- * 
- * Generated by: https://github.com/swagger-api/swagger-codegen.git
- */
-using System;
-using System.Linq;
-using System.IO;
-using System.Text;
-using System.Collections;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.ComponentModel.DataAnnotations;
-using System.Runtime.Serialization;
-using Newtonsoft.Json;
-
-namespace IO.Swagger.V1RC03.Models
-{ 
-    /// <summary>
-    /// 
-    /// </summary>
-    [DataContract]
-    public partial class PackageDescription : IEquatable<PackageDescription>
-    { 
-        /// <summary>
-        /// Gets or Sets AasIds
-        /// </summary>
-
-        [DataMember(Name="aasIds")]
-        public List<string> AasIds { get; set; }
-
-        /// <summary>
-        /// Gets or Sets PackageId
-        /// </summary>
-
-        [DataMember(Name="packageId")]
-        public string PackageId { get; set; }
-
-        /// <summary>
-        /// Returns the string presentation of the object
-        /// </summary>
-        /// <returns>String presentation of the object</returns>
-        public override string ToString()
-        {
-            var sb = new StringBuilder();
-            sb.Append("class PackageDescription {\n");
-            sb.Append("  AasIds: ").Append(AasIds).Append("\n");
-            sb.Append("  PackageId: ").Append(PackageId).Append("\n");
-            sb.Append("}\n");
-            return sb.ToString();
-        }
-
-        /// <summary>
-        /// Returns the JSON string presentation of the object
-        /// </summary>
-        /// <returns>JSON string presentation of the object</returns>
-        public string ToJson()
-        {
-            return JsonConvert.SerializeObject(this, Formatting.Indented);
-        }
-
-        /// <summary>
-        /// Returns true if objects are equal
-        /// </summary>
-        /// <param name="obj">Object to be compared</param>
-        /// <returns>Boolean</returns>
-        public override bool Equals(object obj)
-        {
-            if (ReferenceEquals(null, obj)) return false;
-            if (ReferenceEquals(this, obj)) return true;
-            return obj.GetType() == GetType() && Equals((PackageDescription)obj);
-        }
-
-        /// <summary>
-        /// Returns true if PackageDescription instances are equal
-        /// </summary>
-        /// <param name="other">Instance of PackageDescription to be compared</param>
-        /// <returns>Boolean</returns>
-        public bool Equals(PackageDescription other)
-        {
-            if (ReferenceEquals(null, other)) return false;
-            if (ReferenceEquals(this, other)) return true;
-
-            return 
-                (
-                    AasIds == other.AasIds ||
-                    AasIds != null &&
-                    AasIds.SequenceEqual(other.AasIds)
-                ) && 
-                (
-                    PackageId == other.PackageId ||
-                    PackageId != null &&
-                    PackageId.Equals(other.PackageId)
-                );
-        }
-
-        /// <summary>
-        /// Gets the hash code
-        /// </summary>
-        /// <returns>Hash code</returns>
-        public override int GetHashCode()
-        {
-            unchecked // Overflow is fine, just wrap
-            {
-                var hashCode = 41;
-                // Suitable nullity checks etc, of course :)
-                    if (AasIds != null)
-                    hashCode = hashCode * 59 + AasIds.GetHashCode();
-                    if (PackageId != null)
-                    hashCode = hashCode * 59 + PackageId.GetHashCode();
-                return hashCode;
-            }
-        }
-
-        #region Operators
-        #pragma warning disable 1591
-
-        public static bool operator ==(PackageDescription left, PackageDescription right)
-        {
-            return Equals(left, right);
-        }
-
-        public static bool operator !=(PackageDescription left, PackageDescription right)
-        {
-            return !Equals(left, right);
-        }
-
-        #pragma warning restore 1591
-        #endregion Operators
-    }
-}
